update to 1.2.0

+++ adbyby/adbyby/adbyby/adbyby.sh
@@ -129,13 +129,21 @@ flush_nat(){
 }
 
 creat_start_up(){
-	[ ! -L "/jffs/softcenter/init.d/S93Adbyby.sh" ] && {
-		ln -sf /jffs/softcenter/adbyby/adbyby.sh /jffs/softcenter/init.d/S93Adbyby.sh
-	}
+	if [ "$(nvram get productid)" = "BLUECAVE" ];then
+		[ ! -f "/jffs/softcenter/init.d/M93Adbyby.sh" ] && cp -r /jffs/softcenter/adbyby/adbyby.sh /jffs/softcenter/init.d/M93Adbyby.sh
+	else
+		[ ! -L "/jffs/softcenter/init.d/S93Adbyby.sh" ] && ln -sf /jffs/softcenter/adbyby/adbyby.sh /jffs/softcenter/init.d/S93Adbyby.sh
+	fi
+	if [ "$adbyby_enable" == "0" ];then
+		if [ "$(nvram get productid)" = "BLUECAVE" ];then
+			rm -rf /jffs/softcenter/init.d/M93Adbyby.sh
+		else
+			rm -rf /jffs/softcenter/init.d/S93Adbyby.sh
+		fi
+	fi
 }
 
-
-case $ACTION in
+case $1 in
 start)
 	set_lock
 	if [ "$adbyby_enable" == "1" ];then
